+++ src/core/com/frostwire/search/extratorrent/ExtratorrentSearchPerformer.java
@@ -33,7 +33,7 @@ import com.frostwire.websearch.TorrentWebSearchResult;
  */
 public class ExtratorrentSearchPerformer extends PagedWebSearchPerformer {
 
-    public ExtratorrentSearchPerformer(int token, String keywords, int timeout) {
+    public ExtratorrentSearchPerformer(long token, String keywords, int timeout) {
         super(token, keywords, timeout, 1);
     }
