+++ src/core/com/frostwire/search/mininova/MininovaSearchPerformer.java
@@ -33,7 +33,7 @@ import com.frostwire.websearch.TorrentWebSearchResult;
  */
 public class MininovaSearchPerformer extends PagedWebSearchPerformer {
 
-    public MininovaSearchPerformer(int token, String keywords, int timeout) {
+    public MininovaSearchPerformer(long token, String keywords, int timeout) {
         super(token, keywords, timeout, 1);
     }
