+++ src/core/com/frostwire/search/vertor/VertorSearchPerformer.java
@@ -33,7 +33,7 @@ import com.frostwire.websearch.TorrentWebSearchResult;
  */
 public class VertorSearchPerformer extends PagedWebSearchPerformer {
 
-    public VertorSearchPerformer(int token, String keywords, int timeout) {
+    public VertorSearchPerformer(long token, String keywords, int timeout) {
         super(token, keywords, timeout, 1);
     }
